Feat profile page (#33)
* fix: fix crashes on xiaomi device

* fix: fix crashes on xiaomi device

* update: update version to 1.7.7

* fix: fix auto sign in

* update: update version to 1.7.8

* feat: add jrs ball videos

* fix: add a delay to avoid user info saving failure

* feat: support show image in notification page

* fix: fix jsoup io exception

* fix: revert jsoup upgrade

* fix: fix auto sign-in

* fix: temporary remove login-check action

* update: add rounded corner for all avatars and add profile page

* clean: code clean

* feat: add profile popup dialog

* feat: add follow support

* feat: add follow page

* fix: fix avatar transformation

* fix: fix fetch data npe

* fix: fix load more logic

* update: update version to 1.8.0

* update: update version

* fix: fix crash #52411 Receiver not registered: android.widget.ZoomButtonsController

* fix: fix socket time out crash when app init

* fix: could not find Fragment constructor

* update: update dependencies

* update: add child root on me tab

* change: change code structure and add a webview jump for favorites

* fix: fix toolbar display issue

* fix: fix wrong fragment tag logic

* fix: temporary fix forum page

* fix: fix wrong fragment manager

* update: update provider

* feat: add swipe feature for recycler view

* fix: reduce login check range

* fix: fix lazy refresh on me fragment

* stat: add event for profile page

* fix: add a way to resolve anti scraping mechanism

* stat: add follow page stat

* feat: add follow feed fragment

* update: optimize imports

* update: optimize html text display

* stat: add open preview stat

* fix: add permission check

* feat: add jump to original article for follow feeds

* feat: display image for following feeds

* fix: code cleanup

* change: change following feed title

* change: change follow page title

* feat: support open profile page in article page

* fix: fix private message bubble display issue

* stat: add open profile link event

* stat: add article capture and share stat

* update: update version to 1.8.1

+++ framework/src/main/java/top/easelink/framework/guard/AppGuard.kt
@@ -5,6 +5,7 @@ import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Process
 import timber.log.Timber
+import top.easelink.framework.BuildConfig
 
 class AppGuard(
     private var mContext: Context,
@@ -12,9 +13,11 @@ class AppGuard(
 ): Thread.UncaughtExceptionHandler{
 
     override fun uncaughtException(t: Thread?, e: Throwable?) {
-        restartApp(mContext)
-        mDefaultUncaughtExceptionHandler?.uncaughtException(t, e)
-        Process.killProcess(Process.myPid())
+        if (!BuildConfig.DEBUG) {
+            restartApp(mContext)
+            mDefaultUncaughtExceptionHandler?.uncaughtException(t, e)
+            Process.killProcess(Process.myPid())
+        }
     }
 
     private fun restartApp(context: Context) {
